split authenticator, move to whyis package
