+++ whyis/authenticator/authenticator.py
@@ -1,0 +1,3 @@
+class Authenticator:
+    def authenticate(self, request, datastore, config):
+        pass
